# api/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Device:
    properties:
      connect_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceConnectType'
        description: 设备连接类型 inside内置 / outside外置
      created_at:
        description: 创建时间
        type: string
      device_brand:
        description: 设备品牌
        type: string
      device_code:
        description: 设备编号/序列号
        type: string
      device_model:
        description: 设备型号
        type: string
      device_stall_print_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceStallPrintType'
        description: 出品部门总分单
      device_stall_receipt_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceStallReceiptType'
        description: 打印出品部门全部票据
      device_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceType'
        description: 设备类型
      dining_ways:
        description: 订单类型/就餐方式
        items:
          $ref: '#/definitions/domain.DiningWay'
        type: array
      enabled:
        description: 启用/停用状态
        type: boolean
      id:
        description: 设备 ID
        type: string
      ip:
        description: 设备 IP 地址
        type: string
      location:
        allOf:
        - $ref: '#/definitions/domain.DeviceLocation'
        description: 设备位置
      merchant_id:
        description: 所属商户 ID
        type: string
      name:
        description: 设备名称
        type: string
      open_cash_drawer:
        description: 收银机 开钱箱
        type: boolean
      order_channels:
        description: 订单来源
        items:
          $ref: '#/definitions/domain.OrderChannel'
        type: array
      paper_size:
        allOf:
        - $ref: '#/definitions/domain.PaperSize'
        description: 打印纸张尺寸
      sort_order:
        description: 排序
        type: integer
      stall_id:
        description: 出品部门 ID
        type: string
      stall_name:
        description: 出品部门名称
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.DeviceStatus'
        description: 设备状态
      store_id:
        description: 所属门店 ID
        type: string
      store_name:
        description: 门店名称
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  domain.DeviceConnectType:
    enum:
    - inside
    - outside
    type: string
    x-enum-comments:
      DeviceConnectTypeInside: 内置
      DeviceConnectTypeOutside: 外置
    x-enum-varnames:
    - DeviceConnectTypeInside
    - DeviceConnectTypeOutside
  domain.DeviceLocation:
    enum:
    - front_hall
    - back_kitchen
    type: string
    x-enum-comments:
      DeviceLocationBackKitchen: 后厨
      DeviceLocationFrontHall: 前厅
    x-enum-varnames:
    - DeviceLocationFrontHall
    - DeviceLocationBackKitchen
  domain.DeviceStallPrintType:
    enum:
    - all
    - combined
    - separate
    type: string
    x-enum-comments:
      DeviceStallPrintTypeAll: 全部打印 总单+分单
      DeviceStallPrintTypeCombined: 总单
      DeviceStallPrintTypeSeparate: 分单
    x-enum-varnames:
    - DeviceStallPrintTypeAll
    - DeviceStallPrintTypeCombined
    - DeviceStallPrintTypeSeparate
  domain.DeviceStallReceiptType:
    enum:
    - all
    - exclude
    type: string
    x-enum-comments:
      DeviceStallReceiptTypeAll: 全部打印
      DeviceStallReceiptTypeExclude: 剔除部门商户
    x-enum-varnames:
    - DeviceStallReceiptTypeAll
    - DeviceStallReceiptTypeExclude
  domain.DeviceStatus:
    enum:
    - online
    - offline
    type: string
    x-enum-comments:
      DeviceStatusOffline: 离线
      DeviceStatusOnline: 在线
    x-enum-varnames:
    - DeviceStatusOnline
    - DeviceStatusOffline
  domain.DeviceType:
    enum:
    - cashier
    - printer
    type: string
    x-enum-comments:
      DeviceTypeCashier: 收银机
      DeviceTypePrinter: 打印机
    x-enum-varnames:
    - DeviceTypeCashier
    - DeviceTypePrinter
...
  domain.DiningWay:
    enum:
    - dine_in
    - take_out
    - delivery
    type: string
    x-enum-comments:
      DiningWayDelivery: 外送
      DiningWayDineIn: 堂食
      DiningWayTakeOut: 外带
    x-enum-varnames:
    - DiningWayDineIn
    - DiningWayTakeOut
    - DiningWayDelivery
...
  domain.OrderChannel:
    enum:
    - pos
    - self_order
    - mini_program
    - mobile_order
    - scan_order
    - third_delivery
    type: string
    x-enum-comments:
      OrderChannelMiniProgram: 小程序
      OrderChannelMobileOrder: 手机点餐
      OrderChannelPOS: POS 端
      OrderChannelScanOrder: 扫码点餐
      OrderChannelSelfOrder: 自助点餐
      OrderChannelThirdDelivery: 三方外卖
    x-enum-varnames:
    - OrderChannelPOS
    - OrderChannelSelfOrder
    - OrderChannelMiniProgram
    - OrderChannelMobileOrder
    - OrderChannelScanOrder
    - OrderChannelThirdDelivery
  domain.PaperSize:
    enum:
    - 58mm
    - 80mm
    type: string
    x-enum-varnames:
    - PaperSize58mm
    - PaperSize80mm
...
  response.Response:
    properties:
      code:
        type: string
      data: {}
    type: object
...
  types.DeviceCashier:
    properties:
      open_cash_drawer:
        description: 开启钱箱
        type: boolean
    type: object
  types.DeviceCreateReq:
    properties:
      device_brand:
        description: 设备品牌
        maxLength: 100
        type: string
      device_cashier:
        allOf:
        - $ref: '#/definitions/types.DeviceCashier'
        description: 收银机设备配置
      device_code:
        description: 设备编号
        maxLength: 100
        type: string
      device_model:
        description: 设备型号
        maxLength: 100
        type: string
      device_print:
        allOf:
        - $ref: '#/definitions/types.DevicePrint'
        description: 打印机设备配置
      device_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceType'
        description: 设备类型
        enum:
        - cashier
        - printer
      enabled:
        description: 是否启用
        type: boolean
      location:
        allOf:
        - $ref: '#/definitions/domain.DeviceLocation'
        description: 设备位置
        enum:
        - front_hall
        - back_kitchen
      name:
        description: 设备名称
        maxLength: 50
        type: string
      sort_order:
        description: 排序值，越小越靠前
        minimum: 0
        type: integer
      store_id:
        description: 门店 ID
        type: string
    required:
    - device_code
    - device_type
    - location
    - name
    - store_id
    type: object
  types.DeviceListResp:
    properties:
      devices:
        description: 设备列表
        items:
          $ref: '#/definitions/domain.Device'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  types.DevicePrint:
    properties:
      connect_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceConnectType'
        description: 设备连接类型 inside内置 / outside外置
        enum:
        - inside
        - outside
      device_stall_print_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceStallPrintType'
        description: 打印出品部门总分单
        enum:
        - all
        - combined
        - separate
      device_stall_receipt_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceStallReceiptType'
        description: 打印出品部门全部票据
        enum:
        - all
        - exclude
      dining_ways:
        description: 用餐方式
        items:
          $ref: '#/definitions/domain.DiningWay'
        type: array
      ip:
        description: 打印机 IP 地址
        maxLength: 50
        type: string
      order_channels:
        description: 订单来源
        items:
          $ref: '#/definitions/domain.OrderChannel'
        type: array
      paper_size:
        allOf:
        - $ref: '#/definitions/domain.PaperSize'
        description: 纸张大小
        enum:
        - 58mm
        - 80mm
      stall_id:
        description: 出品部门 ID
        type: string
    required:
    - connect_type
    - device_stall_print_type
    - device_stall_receipt_type
    - dining_ways
    - ip
    - order_channels
    - paper_size
    - stall_id
    type: object
  types.DeviceUpdateReq:
    properties:
      device_brand:
        description: 设备品牌
        maxLength: 100
        type: string
      device_cashier:
        allOf:
        - $ref: '#/definitions/types.DeviceCashier'
        description: 收银机设备配置
      device_code:
        description: 设备编号
        maxLength: 100
        type: string
      device_model:
        description: 设备型号
        maxLength: 100
        type: string
      device_print:
        allOf:
        - $ref: '#/definitions/types.DevicePrint'
        description: 打印机设备配置
      device_type:
        allOf:
        - $ref: '#/definitions/domain.DeviceType'
        description: 设备类型
        enum:
        - cashier
        - printer
      enabled:
        description: 是否启用
        type: boolean
      location:
        allOf:
        - $ref: '#/definitions/domain.DeviceLocation'
        description: 设备位置
        enum:
        - front_hall
        - back_kitchen
      name:
        description: 设备名称
        maxLength: 50
        type: string
      sort_order:
        description: 排序值，越小越靠前
        minimum: 0
        type: integer
      store_id:
        description: 门店 ID
        type: string
    required:
    - device_type
    - location
    - name
    - store_id
    type: object
...
info:
  contact: {}
  description: 供品牌商后台调用.
  title: 品牌商后台 API
  version: "1.0"
paths:
...
  /restaurant/device:
    get:
      description: 分页查询设备列表
      parameters:
      - description: 设备类型
        enum:
        - cashier
        - printer
        in: query
        name: device_type
        type: string
        x-enum-comments:
          DeviceTypeCashier: 收银机
          DeviceTypePrinter: 打印机
        x-enum-varnames:
        - DeviceTypeCashier
        - DeviceTypePrinter
      - description: 设备名称模糊查询
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      - description: 设备状态
        enum:
        - online
        - offline
        in: query
        name: status
        type: string
        x-enum-comments:
          DeviceStatusOffline: 离线
          DeviceStatusOnline: 在线
        x-enum-varnames:
        - DeviceStatusOnline
        - DeviceStatusOffline
      - description: 门店 ID
        in: query
        name: store_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.DeviceListResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取设备列表
      tags:
      - 设备管理
    post:
      consumes:
      - application/json
      description: 创建设备
      parameters:
      - description: 请求信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.DeviceCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: 创建设备
      tags:
      - 设备管理
  /restaurant/device/{id}:
    delete:
      description: 删除设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: No Content
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: 删除设备
      tags:
      - 设备管理
    get:
      description: 根据设备ID获取详情
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Device'
              type: object
      security:
      - BearerAuth: []
      summary: 获取设备详情
      tags:
      - 设备管理
    put:
      consumes:
      - application/json
      description: 更新设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      - description: 请求信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.DeviceUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 更新设备
      tags:
      - 设备管理
  /restaurant/device/{id}/disable:
    put:
      description: 将设备置为禁用
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 禁用设备
      tags:
      - 设备管理
  /restaurant/device/{id}/enable:
    put:
      description: 将设备置为启用
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 启用设备
      tags:
      - 设备管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"